### 2020-02-07 #2078 @whedon
[PRE REVIEW]: LATTE: Lightcurve Analysis Tool for Transiting Exoplanets
**Submitting author:** @noraeisner (<a href="http://orcid.org/0000-0002-9138-9028">N. L. Eisner</a>)
**Repository:** <a href="https://github.com/noraeisner/LATTE" target ="_blank">https://github.com/noraeisner/LATTE</a>
**Version:** v0.0.12
