### 2020-02-07 #2078 @arfon
:wave: @noraeisner. Thanks for your submission to JOSS!

Can you tell me if this paper has an accompanying paper with AAS journals?
