### 2020-02-07 #2078 @noraeisner
Hi Arfon! No, this paper doesn't have an accompanying AAS paper.
