+++ README.md
@@ -16,11 +16,12 @@ email: *nora.eisner@new.ox.ac.uk*
 
 --------
 --------
+
 ### Installation
 
-You can download the code directly from github. Alternatively you can install LATTE using pip (https://pypi.org/project/tessLATTE/) via your command line with:
+LATTE requires python3 to be installed on your computer, which can be download from https://www.python.org/downloads/. Once you have python3, you can download the code directly from github. Alternatively you can install LATTE using pip (https://pypi.org/project/tessLATTE/) via your command line with:
 
-	pip install tessLATTE      
+	pip3 install tessLATTE      
 
 In order for LATTE to work you will need to have the right versions of certain modules installed, so downloading it in a virtual environemt. **Note: ensure that the matplotlib version that you are using is v3.2.0rc1 (pip install matplotlib==3.2.0rc1).**
 
@@ -31,7 +32,7 @@ The first time that the program is run you will be prompted to enter a path to a
 
 LATTE is simply run through the command line with:
 
-	python3 -m LATTE      or     python3 -m LATTE --new-data     (if run for the first time)
+	python3 -m LATTE      or     python3 -m LATTE --new-data     (if run for the first time or when new data is released)
 
 This will open up a box for you that prompts you to enter a TIC ID and indicate whether you would like to extract the information from the 2-minute cadence ('Standard mode') or 30-minute candence Full Frame Image (FFI) data ('FFI mode').
 
@@ -49,8 +50,11 @@ In *FFI mode* the data is downloaded using TESScut and the data detrended using 
 
 ## Input target list
 
-In order to efficiently generate diagnostic plots for multiple targets without having to interactively enter the target TIC IDs, ``LATTE`` can be executed with an input file that lists the TIC IDs, the times of the transit-like events, and the observational sectors to consider (if known). See example. In the cases where the times of the transit events or the sectors have not been entered, the user is prompted to enter these manually via the GUI, as descibed above. For longer target lists the code can also be parallelized (see example).
+In order to efficiently generate diagnostic plots for multiple targets without having to interactively enter the target TIC IDs, ``LATTE`` can be executed with an input file that lists the TIC IDs, the times of the transit-like events, and the observational sectors to consider (if known). See example. In the cases where the times of the transit events or the sectors have not been entered, the user is prompted to enter these manually via the GUI, as descibed above. For longer target lists the code can also be parallelized (see example). If you want to run it with this mode, enter: 
 
+	python3 -m LATTE --targetlist*=path/to/the/csv/input/file*
+
+The program will skip targets that have already been analysed by you, so if you want to overwrite data add '--o' to the command.
 
 ### Transit time selection
 
@@ -138,41 +142,29 @@ The code will then generate download and process all of the data. Note that all 
 
 ### Arguments
 
-NOTE: all of these arguments (except new-path, auto and targetlist) can be changed as option in the GUI. They are arguments in case the same options wish to be run multiple times and the user therfore wishes to identify them in the command line when the program is executed.
-
+NOTE: all of these arguments (except new-path, auto and targetlist, new-data) can be changed as option in the GUI. They are arguments in case the same options wish to be run multiple times and the user therfore wishes to identify them in the command line when the program is executed.
 
 !!!!!!  **--new-data**  The code requires multiple text files to be stored on your computer in order to run - these are downloaded automatically from the MAST server. The first time the proghram is run, and any time that there is new data available, add **--new-data** to the command line when running the program. The code checks what data has already been downloaded and doesn't re-download anything that already exists.
 
-**--tic** You can skip the box to enter the tic ID by entering it in the command line with e.g. --tic=55525572. 
-
-**--sector** You can skip entering the sectors by entering them in the command line with e.g. --sector=2,5. You will need to know in what sectors this target was observed.
[…]
