### 2020-02-13 #2078 @xuanxu
@christinahedges Thanks! If you don't have previous/recent association with the author and you think you can make an impartial scientific evaluation I don't think there's conflict, or if there is it's minor and can be waived.
