### 2020-02-13 #2101 @whedon
[REVIEW]: LATTE: Lightcurve Analysis Tool for Transiting Exoplanets
**Submitting author:** @noraeisner (<a href="http://orcid.org/0000-0002-9138-9028">N. L. Eisner</a>)
**Repository:** <a href="https://github.com/noraeisner/LATTE" target ="_blank">https://github.com/noraeisner/LATTE</a>
**Version:** v1.0.0
