### 2020-02-13 #2101 @xuanxu
@christinahedges, @aureliocarnero, @noraeisner: this is the review thread for the paper. All of our communications will happen here from now on.

Please feel free to ask me (or the submitting author, or each other) if you have any questions :)
