### 2020-02-13 #2101 @xuanxu
@danielskatz Agreed. By her comment I understood that she is not collaborating yet and I assumed there was time enough for the review to finish before the visit, but you are right that's just an assumption I made. @christinahedges can you please clarify this point?
