### 2020-02-13 #2101 @christinahedges
Hi @xuanxu and @danielskatz, yes I am not currently collaborating with Norah, our collaboration starts in a few months. If you're happy this is not a conflict then I'm happy! 👍
