### 2020-02-16 #2101 @aureliocarnero
Dear @noraeisner, I'm reviewing the installation instructions, and following your recommendations, I came to a message error telling me that:

ModuleNotFoundError: No module named 'tkinter'

It was easy to solve it, I need to do a: "sudo apt-get install python3-tk", but I think it would be nice to explicitly say that in the documentation, same as you name with the matplotlib version: pip install matplotlib==3.2.0rc1
Put another note on tkinter.
