### 2020-02-16 #2101 @aureliocarnero
Also, in the installation instructions, when you say that "The first time that the code is run it will also have to download the text data files from MAST (this only has to run once)."
Could you please state the size of the data that will be download? Which is ~325M
Thanks  @noraeisner
