### 2020-02-17 #2101 @noraeisner
Dear @aureliocarnero, thank you so much for reviewing the code. I will definitely add the need for tkinter and the size of dowloaded data to the installation instructions. Thank you so much.
