### 2020-02-19 #2101 @aureliocarnero
Dear @noraeisner, there is a few comments to make.
In order to validate the documentation, you should explain how to perform the automated test. I see there is a directory called tests. Could you put in the code documentation how to run the tests? Ideally, should be done in automatic mode, like using pytests.
