### 2020-02-19 #2101 @aureliocarnero
@noraeisner another thing is documentation is lacking Community guidelines, please put those (see above in the checklist about what it means). If you don't know what these are, you can see examples in other JOSS code repositories
