### 2020-02-19 #2101 @aureliocarnero
@noraeisner

Also in the README.md file, there are some exclamation marks that I think are an error
"""""""""""""""""""""""""""""""
Arguments

NOTE: all of these arguments (except new-path, auto and targetlist, new-data) can be changed as option in the GUI. They are arguments in case the same options wish to be run multiple times and the user therfore wishes to identify them in the command line when the program is executed.

!!!!!! --new-data
""""""""""""""""""""""""""""""""""""""""
There, before --new-data

Also, correct the word therefore (it is written therfore)
