### 2020-02-19 #2101 @aureliocarnero
@noraeisner also there is now a new setting called model that you do not explain what it is.
Could you please document it?
Cheers!
