### 2020-02-19 #2101 @aureliocarnero
@noraeisner I'm playing around with the code and there is something I would recommend that you improve.
When I run the diagnostic plots, after clicking DONE, then the GUI closes, but the code still is working in the background, without noticing. Therefore, it is a bit confusing because the user might think the process is finished, when it is not. It still keeps running in the background without notice
