### 2020-02-20 #2101 @aureliocarnero
Hello, @noraeisner, I opened an issue in the github repo because I found a bug so I think it is better it is registered there.
Cheers
Aurelio
