+++ paper/paper.md
@@ -104,7 +104,7 @@ The next release will allow for the option to model the transit-like events usin
 
 # Acknowledgements
 
-I thank all of the Planet Hunters *TESS* volunteers whose dedication to the project encouraged me to write this analysis tool. I am also extremely grateful to all of the support provided by the Zooniverse team and the Oxford exoplanet group.
+We thank all of the Planet Hunters *TESS* volunteers whose dedication to the project encouraged me to write this analysis tool. We are also extremely grateful to all of the support provided by the Zooniverse team and the Oxford exoplanet group.
 
 
 # References
