+++ CONTRIBUTING.md
@@ -1,0 +1,48 @@
+
+# Contributing to LATTE
+
+-----------
+
+Thank you so much for using and contributing to LATTE! Please read the guidlines below before contributing.
+
+
+### Reporting a bug
+
+If the code is not working as descibed you may have probably discovered a bug! If this is the case we will work together to try and fix it as quckly as possible. Please alert us to the bug by submitting an issue to the LATTE github [issue](https://github.com/noraeisner/LATTE/issues) page:
+
+When reporting an issue please include the following:  
+
+- Clear title and short deciption of the issue. 
+- A reproducible example of the issue including the output or error message.
+- Briefly compare the expected behavior to the actual output. 
+
+Before opening an issues, check that the bug that you are reporting is not an already open issue by searching the GitHub issues. 
+
+### Adding new features
+
+If you have not come across a bug but feel that there are features missing from the code that you would like to see implemented, I invite you to: 
+
+- *Fork the repository and develop your own additions*. New code should be well documented and be testable with unittests. Please also clearly describe the new feature, and how it improves LATTE. 
+
+- *Contibuting ideas*. If you would like to see features included but don't know how to code them, please contact the authors directly (via email) and we can discuss the possibility of new tools. 
+
+
+All contrbutions and ideas are most welcome! If you have any queries that aren't a bug please do not open a GitHUb issue but contact the authors directly with your query. 
+
+Thank you! 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
