+++ LATTE/pyaneti_LATTE/src/plot_rv.py
@@ -1,227 +1,0 @@
-#===========================================================
-#                   RV PLOTS
-#===========================================================
-#===========================================================
-#        Plot RV timeseries
-#===========================================================
-def plot_rv_timeseries():
-#-------------------------------------------------------------------------------
-    #Convert factor from km to m
-    cfactor = np.float(1.e3)
-    data_rv_inst = [None]*nt
-    data_erv_inst = [None]*nt
-    for i in range(0,nt):
-      data_rv_inst[i] = list(rv_all[i])
-      data_erv_inst[i] = list(errs_all[i])
-#-------------------------------------------------------------------------------
-    #CREATE MODEL TO PLOT
-    #MODEL LENGHT IS THE WHOLE RANGE +/- 10%
-    xmin = min(np.concatenate(time_all))
-    xmax = max(np.concatenate(time_all))
-    total_tt = int(xmax - xmin)
-    add = total_tt*0.1
-    xmax = xmax + add
-    xmin = xmin - add
-    n = total_tt*100
-    if ( n > 50000): n = 50000
-    #CREATE TIME VECTOR
-    rvx = np.arange(xmin,xmax,(xmax-xmin)/n)
-    #COMPUTE MODEL WITH ALL THE PLANETS
-    rvy = pti.rv_curve_mp(rvx,0.0,t0_val,k_val,P_val,e_val,w_val,alpha_val,beta_val)
-#-------------------------------------------------------------------------------
-    #CORRECT DATA FOR EACH SPECTROGRAPH OFFSET
-    rv_no_offset = [None]*nt
-    rv_residuals = [None]*nt
-    for j in range(0,nt):
-      #Remove the compute offset for instrument j
-      rv_no_offset[j] = np.asarray(data_rv_inst[j] - v_val[j])
-      #Store the model with the planets to compute the residuals for instrument j
-      rv_residuals[j] = pti.rv_curve_mp(time_all[j],0.0,t0_val,\
-      k_val,P_val,e_val,w_val,alpha_val,beta_val)
-      #Compute the residuals for the instrument j
-      rv_residuals[j] = np.asarray(rv_no_offset[j] - rv_residuals[j])
-#-------------------------------------------------------------------------------
-    if kernel_rv[0:2] == 'VR' or kernel_rv[0:3] == 'R15' or kernel_rv[0:2] == 'MQ':
-        #How many timeseries?
-        ns = int((len(fit_krv) - 3)/2)
-        #This vector contains all the timeseries (a TxN vector)
-        xvec = mega_time
-        #This vector contains the predicted timeseries for N cases
-        rvx_tot = np.concatenate([rvx]*ns)
-        #Let us create our vector with the residuals for the N timeseries
-        yvec = [None]*len(mega_rv)
-        #Our first chunk of timeseries is always the RV, in this case, we have to
-        #Create the residuals once removed the planet signals
-        yvec_noplanet = np.concatenate(rv_no_offset)
-        yvec_planet = np.concatenate(rv_residuals)
-        #yvec_planet = np.concatenate(rv_no_offset)
-        for i in range(int(len(mega_rv)/ns)):
-          yvec[i] = yvec_planet[i]
-        #Now, let us store the ancilliary data vectors
[…]
