+++ tessLATTE.egg-info/SOURCES.txt
@@ -1,0 +1,26 @@
+MANIFEST.in
+README.md
+setup.py
+LATTE/LATTE_DV.py
+LATTE/LATTEbrew.py
+LATTE/LATTEutils.py
+LATTE/__init__.py
+LATTE/__main__.py
+LATTE/filters.py
+LATTE/norm.py
+LATTE/pyaneti_LATTE.py
+LATTE/LATTE_imgs/.DS_Store
+LATTE/LATTE_imgs/LATTE_logo.jpeg
+LATTE/LATTE_imgs/LATTE_logo.pdf
+LATTE/LATTE_imgs/LATTE_logo.png
+LATTE/LATTE_imgs/LATTE_logo_old.jpg
+LATTE/LATTE_imgs/LATTE_logo_small.png
+LATTE/LATTE_imgs/PHT_logo.jpg
+LATTE/LATTE_imgs/TESS_logo.png
+LATTE/pyaneti_LATTE/__init__.py
+tessLATTE.egg-info/PKG-INFO
+tessLATTE.egg-info/SOURCES.txt
+tessLATTE.egg-info/dependency_links.txt
+tessLATTE.egg-info/entry_points.txt
+tessLATTE.egg-info/requires.txt
+tessLATTE.egg-info/top_level.txt
