### 2020-02-20 #2101 @noraeisner
Hi @aureliocarnero, thank you so much for all your comments. I have added a section in the readme file on how to execute the automated tests (using python3 -m unittest) and I have added a section and a file (CONTRIBUTIONS.md) for the Community guidelines as well as addressed the exclamation marks and spell mistakes in the readme file. I'm so sorry about missing these.

The modelling feature is unfortunately not quite ready to be distributed as I'm finding that it struggles to run on certain computers and is compiler specific so I have removed the 'model' option entirely until this is better tested.

I haven't been able to reproduce the error of the 'done' button freezing yet but I am currently looking into it. I'm also trying to reproduce the  'self.canvas.get_property("window").set_cursor(cursord[cursor])'  error. Can I ask what operating software you are using so that I can try and track these issues down that way? I have tested the code on Mac and Linux/Fedora 27 only.

Thank you!
