### 2020-02-20 #2101 @aureliocarnero
Thank you a lot @noraeisner for the quick response.
So I cannot reproduce again either the canvas.get_property("window") error any longer! It remains a mystery why I had 2 separate errors in the same window of time...
I'm using Ubuntu 19.10

I will test again to see if I get an error again.

To try the new version where the settings "model" is no longer available... Do I have to install the code again, I guess?
