### 2020-02-20 #2101 @aureliocarnero
@noraeisner two quick comments:
1) Could you please put in the paper, a reference to Kepler? Second paragraph in the paper, I'm missing the reference to Kepler there.

2) Maybe it is a silly thing, May I would recommend to put the Brasseur, Phillip, Fleming, Mullally & White, 2019 as Brasseur et al, 2019.?

3) Please correct the word "eecuting" in the README.md, here:
To check, please try eecuting the following command in from your command line: python -m tess_stars2px -t 261136679
