+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 NASA's Transiting Exoplanet Survey Satellite [*TESS*, @ricker15] is the first nearly all-sky space-based exoplanet search mission. During the nominal two year primary mission alone, *TESS* will monitor over 80 per cent of the sky with the prime objective of  discovering transiting exoplanets around 200,000 pre-selected bright nearby stars, which are being monitored every two minutes. In addition to these short-cadence observations, *TESS* is obtaining Full Frame Images (FFIs) every thirty minutes, which include brightness measurements of millions of additional stars. 
 
-Unlike its predecessor *Kepler*, *TESS* uses very large pixels, meaning that astrophysical false positives (e.g. where the signal of a target is contaminated by a faint eclipsing binary) are expected to be common. Many false positive scenarios can be ruled out from detailed examination of the *TESS* data using standard diagnostic tests. 
+Unlike its predecessor *Kepler* [@borucki2010kepler], *TESS* uses very large pixels, meaning that astrophysical false positives (e.g. where the signal of a target is contaminated by a faint eclipsing binary) are expected to be common. Many false positive scenarios can be ruled out from detailed examination of the *TESS* data using standard diagnostic tests. 
 
 ``LATTE`` is an open source Python package that performs these standard diagnostic tests to help identify, vet and characterise signals in the *TESS* lightcurves in order to weed out instrumental and astrophysical false positives. The program is designed to allow for a fast, in depth analysis of targets that have already been identified as promising candidates by the main *TESS* pipelines or via alternative methods such as citizen science [e.g., @fischer12; @Christiansen2018; @eisner19]. The implemented diagnostic tests are similar to the ‘Data Validation’ steps performed by the Science Processing Operations Center [SPOC; @jenkins16] pipeline. However, while the SPOC pipeline tests are performed on a selection of two-minute cadence *TESS* targets only, ``LATTE`` allows for the analysis of any *TESS* target with a valid TIC ID [@Stassun19], including both the two-minute and thirty-minute cadence targets.
 
