+++ paper/paper.bib
@@ -206,12 +206,23 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{brasseur2019,
-  title={Astrocut: Tools for creating cutouts of TESS images},
-  author={Brasseur, CE and Phillip, Carlita and Fleming, Scott W and Mullally, SE and White, Richard L},
-  journal={Astrophysics Source Code Library},
-  year={2019}
+
+@MISC{brasseur2019,
+       author = {{Brasseur}, C.~E. and {Phillip}, Carlita and {Fleming}, Scott W. and
+         {Mullally}, S.~E. and {White}, Richard L.},
+        title = "{Astrocut: Tools for creating cutouts of TESS images}",
+     keywords = {Software},
+         year = "2019",
+        month = "May",
+          eid = {ascl:1905.007},
+        pages = {ascl:1905.007},
+archivePrefix = {ascl},
+       eprint = {1905.007},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ascl.soft05007B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
 
 @ARTICLE{Stumpe2012,
        author = {{Stumpe}, Martin C. and {Smith}, Jeffrey C. and {Van Cleve}, Jeffrey E. and
@@ -431,3 +442,13 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@article{borucki2010kepler,
+  title={Kepler planet-detection mission: introduction and first results},
+  author={Borucki, William J and Koch, David and Basri, Gibor and Batalha, Natalie and Brown, Timothy and Caldwell, Douglas and Caldwell, John and Christensen-Dalsgaard, J{\o}rgen and Cochran, William D and DeVore, Edna and others},
+  journal={Science},
+  volume={327},
+  number={5968},
+  pages={977--980},
+  year={2010},
+  publisher={American Association for the Advancement of Science}
+}
