### 2020-02-20 #2101 @noraeisner
@aureliocarnero, I'm glad to hear that the problem is no longer occurring but I agree that it is very strange that two errors happened at the same time. I will test the code on an Ubuntu computer in order to see whether the errors are related to the operating system.

Yes, I have removed the modelling routine from the GitHub repository (and I will do the same for the pip install version) so it should now run without it if is is re-installed. I'm so sorry for the inconvenience.

For the other comments:
1. I just added the Kepler reference. That should  definitely not have been missing.
2. I completely agree with you that it should just be Brasseur et al, 2019 but as far as I can tell the compile automatically decides how to write the references. Please do let me know if there is a way for me to change it manually.
3. I have corrected this spelling mistake, thank you for spotting this.

It has also come to my attention that some computers have an SSL license issue when running the code for the first time. I have added a couple of sentences to the readme explaining how to diagnose and solve this if it does occur.
