### 2020-02-20 #2101 @aureliocarnero
> For the other comments:
>
>     2. I completely agree with you that it should just be Brasseur et al, 2019 but as far as I can tell the compile automatically decides how to write the references. Please do let me know if there is a way for me to change it manually.

About this, don't worry then, it must be the reference style of JOSS, they might have the limit to 5 authors before going to et al.
