### 2020-02-20 #2101 @aureliocarnero
@noraeisner could you please let me know when the LATTE installation in the new version is available in pip install. Maybe already is.
