### 2020-02-23 #2101 @aureliocarnero
@noraeisner please update here the documentation changes we have done through the review:
https://pypi.org/project/tessLATTE/
