### 2020-02-23 #2101 @aureliocarnero
@noraeisner I run the unittests and I had this output (says Run 7 tests, with FAILED errors=3)
1) Could you please put in the section: "Perform automated tests" that you must clone the git repository first and then you must be in the repository home directory to run those?

2) Here is the output of running the unittests:

/home/carnero/codes/LATTE/tests/test_BLS.py:57: ResourceWarning: unclosed file <_io.FileIO name='./tests/tic55525572_lc.fits' mode='rb' closefd=True>
  alltime, allflux, allflux_err, all_md, alltimebinned, allfluxbinned, allx1, allx2, ally1, ally2, alltime12, allfbkg, start_sec, end_sec, in_sec, tessmag, teff, srad = utils.download_data(indir,sector, tic, binfac = 5, test = './tests/tic55525572_lc.fits')
/home/carnero/codes/LATTE/LATTE/filters.py:256: RuntimeWarning: invalid value encountered in greater
  out = np.isnan(work) + (myabs(array - work) > (nsig*sigma))
Detrend data prior to BLS analysis... done.
/home/carnero/.local/lib/python3.7/site-packages/matplotlib/backends/backend_gtk3.py:187: DeprecationWarning: Gtk.Widget.set_double_buffered is deprecated
  self.set_double_buffered(True)
/home/carnero/.local/lib/python3.7/site-packages/matplotlib/backends/backend_gtk3.py:343: DeprecationWarning: Gtk.Window.set_wmclass is deprecated
  self.window.set_wmclass("matplotlib", "Matplotlib")
/home/carnero/.local/lib/python3.7/site-packages/matplotlib/backends/backend_gtk3.py:372: DeprecationWarning: Gtk.Widget.size_request is deprecated
  size_request = child.size_request()
.EEE/home/carnero/codes/LATTE/tests/test_plotting.py:59: ResourceWarning: unclosed file <_io.FileIO name='./tests/tic55525572_lc.fits' mode='rb' closefd=True>
  alltime, allflux, allflux_err, all_md, alltimebinned, allfluxbinned, allx1, allx2, ally1, ally2, alltime12, allfbkg, start_sec, end_sec, in_sec, tessmag, teff, srad = utils.download_data(indir,sector, tic, binfac = 5, test = './tests/tic55525572_lc.fits')
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:2979: RuntimeWarning: invalid value encountered in greater
  lkeep = np.isfinite(X1.sum(axis=1)) * (Q==0) * (X1.sum(axis=1)>0)
/home/carnero/codes/LATTE/tests/test_plotting.py:60: ResourceWarning: unclosed file <_io.FileIO name='./tests/tic55525572_tp.fits' mode='rb' closefd=True>
  X1_list, X4_list, oot_list, intr_list, bkg_list, apmask_list, arrshape_list, t_list, T0_list, tpf_filt_list = utils.download_tpf_mast(indir, transit_sec, transit_list, tic, test = './tests/tic55525572_tp.fits')
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:4170: RuntimeWarning: invalid value encountered in less
  mask_dd = (np.array(time_dd) < peak+0.75) & (np.array(time_dd) > peak-0.75)
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:4170: RuntimeWarning: invalid value encountered in greater
  mask_dd = (np.array(time_dd) < peak+0.75) & (np.array(time_dd) > peak-0.75)
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:4184: RuntimeWarning: invalid value encountered in less
  mask_dd = (np.array(time_dd) < peak+0.75) & (np.array(time_dd) > peak-0.75)
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:4184: RuntimeWarning: invalid value encountered in greater
  mask_dd = (np.array(time_dd) < peak+0.75) & (np.array(time_dd) > peak-0.75)
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:4185: RuntimeWarning: invalid value encountered in less
  mask_dd_binned = (np.array(time_dd_binned) < peak+0.75) & (np.array(time_dd_binned) > peak-0.75)
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:4185: RuntimeWarning: invalid value encountered in greater
  mask_dd_binned = (np.array(time_dd_binned) < peak+0.75) & (np.array(time_dd_binned) > peak-0.75)
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:3395: MatplotlibDeprecationWarning: Adding an axes using the same arguments as a previous axes currently reuses the earlier instance.  In a future version, a new instance will always be created and returned.  Meanwhile, this warning can be suppressed, and the future behavior ensured, by passing a unique label to each axes instance.
  plt.subplot(2,gs,g+1) # define the plotting area
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:3418: MatplotlibDeprecationWarning: Adding an axes using the same arguments as a previous axes currently reuses the earlier instance.  In a future version, a new instance will always be created and returned.  Meanwhile, this warning can be suppressed, and the future behavior ensured, by passing a unique label to each axes instance.
  plt.subplot(2,gs,gs+1+g) # define the plotting area
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:3630: RuntimeWarning: invalid value encountered in less
  mask_dd = (np.array(alltime) < peak+2) & (np.array(alltime) > peak-2)
/home/carnero/codes/LATTE/LATTE/LATTEutils.py:3630: RuntimeWarning: invalid value encountered in greater
  mask_dd = (np.array(alltime) < peak+2) & (np.array(alltime) > peak-2)
1   out of    11
2   out of    11
3   out of    11
4   out of    11
5   out of    11
6   out of    11
7   out of    11
8   out of    11
9   out of    11
10   out of    11
11   out of    11

 Calculating the Aperture Mask... done.

Waiting on plot...
/usr/lib/python3.7/unittest/case.py:628: ResourceWarning: unclosed file <_io.FileIO name='./tests/tic55525572_tp.fits' mode='rb' closefd=True>
  testMethod()
.Running Pyaneti modelling - this could take a while so be patient...
Traceback (most recent call last):
  File "/home/carnero/codes/LATTE/LATTE/pyaneti_LATTE.py", line 43, in <module>
    import pyaneti as pti #FORTRAN module
ImportError: bad magic number in 'pyaneti': b'\x03\xf3\r\n'
..
ERROR: test_DVreport (tests.test_DVreport.TestDVreport)
Traceback (most recent call last):
  File "/home/carnero/codes/LATTE/tests/test_DVreport.py", line 61, in test_DVreport
    ldv.LATTE_DV(tic, indir, syspath, transit_list, sectors_all, target_ra, target_dec, tessmag, teff, srad, [0], [0], tpf_corrupt, FFI = False,  bls = False, model = False, mpi = None, test = './tests/')
TypeError: LATTE_DV() missing 1 required positional argument: 'astroquery_corrupt'

ERROR: tests.test_data_download (unittest.loader._FailedTest)
ImportError: Failed to import test module: tests.test_data_download
Traceback (most recent call last):
  File "/usr/lib/python3.7/unittest/loader.py", line 436, in _find_test_path
    module = self._get_module_from_name(name)
  File "/usr/lib/python3.7/unittest/loader.py", line 377, in _get_module_from_name
    __import__(name)
  File "/home/carnero/codes/LATTE/tests/test_data_download.py", line 17, in <module>
    from nose.tools import assert_true
ModuleNotFoundError: No module named 'nose'

ERROR: tests.test_data_products (unittest.loader._FailedTest)
ImportError: Failed to import test module: tests.test_data_products
Traceback (most recent call last):
  File "/usr/lib/python3.7/unittest/loader.py", line 436, in _find_test_path
    module = self._get_module_from_name(name)
  File "/usr/lib/python3.7/unittest/loader.py", line 377, in _get_module_from_name
    __import__(name)
  File "/home/carnero/codes/LATTE/tests/test_data_products.py", line 17, in <module>
    import mock
ModuleNotFoundError: No module named 'mock'

Ran 7 tests in 36.193s

FAILED (errors=3)
