+++ tessLATTE.egg-info/requires.txt
@@ -12,3 +12,4 @@ tess-point
 reproject==0.4
 reportlab
 astropy==3.1.1
+mock>2.0.0
