+++ example_output/tesspoint/55525572_tesspoint.txt
@@ -1,25 +1,0 @@
-# TIC     |   RA      |   Dec     | EclipticLong | EclipticLat | Sector | Camera | Ccd | ColPix | RowPix
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 |  1 | 4 | 1 |  590.538 | 1605.744
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 |  2 | 4 | 1 |  993.092 |  939.091
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 |  3 | 4 | 1 | 1658.765 |  534.155
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 |  4 | 4 | 2 |  271.126 |  484.900
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 |  5 | 4 | 2 |  979.764 |  795.338
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 |  6 | 4 | 2 | 1462.970 | 1389.746
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 |  8 | 4 | 3 |  701.297 | 1352.696
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 |  9 | 4 | 3 | 1183.664 |  792.473
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 10 | 4 | 3 | 1872.402 |  500.453
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 11 | 4 | 4 |  476.186 |  551.994
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 12 | 4 | 4 | 1143.728 |  957.349
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 13 | 4 | 4 | 1546.583 | 1624.319
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 27 | 4 | 4 | 1618.689 | 2026.325
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 28 | 4 | 1 |  635.354 | 1480.513
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 29 | 4 | 1 | 1057.904 |  877.588
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 30 | 4 | 1 | 1740.383 |  510.702
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 31 | 4 | 2 |  379.794 |  507.825
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 32 | 4 | 2 | 1097.024 |  893.773
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 33 | 4 | 2 | 1535.547 | 1567.330
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 34 | 4 | 3 |  533.056 | 1864.897
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 35 | 4 | 3 |  813.211 | 1169.558
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 36 | 4 | 3 | 1357.334 |  681.517
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 37 | 4 | 3 | 2059.031 |  479.347
-055525572 |  72.694043 | -60.905449 |  28.441696 | -80.531668 | 38 | 4 | 4 |  634.178 |  610.759
