### 2020-02-24 #2101 @noraeisner
Hi @aureliocarnero, I'm so sorry that the tests didn't run - there were some files missing from the GitHub page. I have updated it so that it should now run along with the readme file. I should note that the runtime warnings still appear when the tests are run, however these don't stop the code,  or tests, form completing. I have looked into them and they are a common problem with the 'requests' and 'reportlab', modules as these don't close their files properly. I haven't been able to find a solution to avoid these but please do let me know if they are a big problem and I will try and find a work-around.

I have also updated the pip install version - now v0.1.13. Thank you so much.
