### 2020-03-01 #2101 @aureliocarnero
Dear @noraeisner . Could you please revise the pip3 install routine? After you updated to the new version, I cannot install LATTE correctly. It runs ok, all the dependencies installations, but when I try to run the program, it tells me that "No module named LATTE"
If I install it from the github repo with "pip install .", it works ok.
Cheers
Aurelio
