### 2020-03-01 #2101 @aureliocarnero
@xuanxu thank you, I just saw that it is explained above in the instructions. My fault!
Nonetheless, I do not see the new link to the pdf... Maybe it is because it is taking some time to compile?
