### 2020-03-03 #2101 @aureliocarnero
Just in case you missed this message:

Dear @noraeisner . Could you please revise the pip3 install routine? After you updated to the new version, I cannot install LATTE correctly. It runs ok, all the dependencies installations, but when I try to run the program, it tells me that "No module named LATTE"
If I install it from the github repo with "pip install .", it works ok.

Also, I still have 1 error when I run the unittest. Is this normal?
=================
FAIL: test_nn (LATTE.tests.test_data_download.TestNearestNeighbours)
Traceback (most recent call last):
  File "/home/carnero/codes/LATTE/LATTE/tests/test_data_download.py", line 50, in test_nn
    self.assertEqual(output[1], [0.0, 6.795727328666717, 10.622509290130298, 15.63959237521102, 16.881930554329756, 17.79841005439404], "nearest neighbour distances are incorrect")
AssertionError: Lists differ: [0.0,[32 chars]130296, 15.63959237521102, 16.88193055432957, [14 chars]4076] != [0.0,[32 chars]130298, 15.63959237521102, 16.881930554329756,[14 chars]9404]

First differing element 2:
10.622509290130296
10.622509290130298

  [0.0,
   6.795727328666717,
-  10.622509290130296,
?                   ^

+  10.622509290130298,
?                   ^

   15.63959237521102,
-  16.88193055432957,
?                  ^

+  16.881930554329756,
?                 + ^

-  17.798410054394076]
?                  ^^

+  17.79841005439404]
?                  ^
 : nearest neighbour distances are incorrect

Ran 10 tests in 138.972s

FAILED (failures=1)
===================

Cheers
Aurelio
