+++ LATTE/tests/test_data_download.py
@@ -34,8 +34,8 @@ class TestTESSpoint(unittest.TestCase):
 	def test_testpoint(self):
 		output = LATTEutils.tess_point(outdir, '55525572')
 
-		self.assertEqual(output[1], 72.69404300000001, "TESS point RA is not correct")
-		self.assertEqual(output[2], -60.905449, "TESS point Dec is not correct")
+		self.assertAlmostEqual(output[1], 72.69404300000001,places=5,  msg ="TESS point RA is not correct")
+		self.assertAlmostEqual(output[2], -60.905449, places=4, msg ="TESS point Dec is not correct")
 
 
 class TestNearestNeighbours(unittest.TestCase):
@@ -48,8 +48,8 @@ class TestNearestNeighbours(unittest.TestCase):
 
 		self.assertEqual(output[0], [55525572, 55525518, 358806415, 55557836, 55557135, 55522986], "nearest neighbour TIC IDs are incorrect")
 		self.assertEqual(output[1], [0.0, 6.795727328666717, 10.622509290130298, 15.63959237521102, 16.881930554329756, 17.79841005439404], "nearest neighbour distances are incorrect")
-		self.assertEqual(output[2], 72.6941, "RA is incorrect")
-		self.assertEqual(output[3], -60.9055, "DEC is incorrect")
+		self.assertAlmostEqual(output[2], 72.6941, places=3, msg="RA is incorrect")
+		self.assertAlmostEqual(output[3], -60.9055, places=3, msg="DEC is incorrect")
 
 
 class TestDownloadLC(unittest.TestCase):
@@ -64,27 +64,27 @@ class TestDownloadLC(unittest.TestCase):
 		alltime, allflux, allflux_err, all_md, alltimebinned, allfluxbinned, allx1, allx2, ally1, ally2, alltimel2, allfbkg, start_sec, end_sec, in_sec, tessmag, teff, srad = LATTEutils.download_data(outdir, [5], '55525572', binfac = 5)
 
 
-		self.assertEqual(float(alltime[0]), 1437.9785214992496, "alltime is incorrect")
-		self.assertEqual(float(allflux[1000]), 1.0011740922927856, "allflux is incorrect")
-		self.assertEqual(float(allflux_err[1000]), 0.0009966295910999179, "allflux_err is incorrect")
-		self.assertEqual(float(all_md[0]), 1441.0243360822994, "all_md is incorrect")
-		self.assertEqual(float(alltimebinned[0]), 1437.9812992567897, "alltimebinned is incorrect")
-		self.assertEqual(float(allfluxbinned[1000]), 1.000087034702301, "alltimebinned is incorrect")
+		self.assertAlmostEqual(float(alltime[0]), 1437.9785214992496, places=5, msg="alltime is incorrect")
+		self.assertAlmostEqual(float(allflux[1000]), 1.0011740922927856, places=5, msg="allflux is incorrect")
+		self.assertAlmostEqual(float(allflux_err[1000]), 0.0009966295910999179, places=5,msg= "allflux_err is incorrect")
+		self.assertAlmostEqual(float(all_md[0]), 1441.0243360822994, places=5, msg="all_md is incorrect")
+		self.assertAlmostEqual(float(alltimebinned[0]), 1437.9812992567897, places=5, msg="alltimebinned is incorrect")
+		self.assertAlmostEqual(float(allfluxbinned[1000]), 1.000087034702301, places=5, msg="alltimebinned is incorrect")
 		
-		self.assertEqual(float(allx1[0]), -0.006673532877357502, "allx1 is incorrect")
[…]
