### 2020-03-04 #2101 @noraeisner
Hi @aureliocarnero , I am so sorry for missing your previous message. I have re-uploaded the pip install version (v0.1.15). I have verified that it works on two computers (Mac and Ubuntu), however, as far as I can tell the previous version also worked for me so please do let me know if the problem persists. I have also changed the unittests so that the different roundings do not raise an error. Thank you so much for looking at the code!
