### 2020-03-04 #2101 @aureliocarnero
Hello @noraeisner . Now the installation works!
I don't know why, but if I installed it with pip3, it gave me an error, but if I did it with pip, then it worked fine.
Actually, when I installed with pip3, and executed with python3.6, then it run, so it might be a problem with my python 3 version setup.
Basically, if I'm in python 3 enviroment already, running pip install tessLATTE worked perfectly, without need of pip3.
