### 2020-03-04 #2101 @aureliocarnero
@noraeisner in terms of the unittest, there is something funny. If I set the --new-path somewhere outside the LATTE directory, it gives me an error, but if I put the path somewhere within the LATTE directory, it runs ok. So I would recommend you say something like this. Could you try to replicate this?

For example: when it asks me about the new-path and I put: ./LATTE_output (what is recommended), then it worked, but if I put /home/carnero/codes/tempLATTE, then it failed.

Nonetheless, now the unittest gives no errors at all! That is good.
I will test the mpi example case, and if I can run it, then I will validate all the code.
We are close!
Cheers
Aurelio
