### 2020-03-09 #2101 @christinahedges
Hi @xuanxu

> Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations

This link has expired for me, would you be able to tell me how I can get a new one? I'm just going through the checklist now!

Thanks!
C
