### 2020-03-09 noraeisner/LATTE#3 @christinahedges
[JOSS Review] 
Hi @noraeisner

I'm completing your [JOSS review](https://github.com/openjournals/joss-reviews/issues/2101), please excuse the lateness of my review! Please find some comments below. I'll add more as I keep digging!

## Installation

I am going through the installation of your package. I am noticing a few things that I thought I would let you know. These don't get in the way of installing your package but might help installation, these are suggestions but by no means stop my reviewing/installing your package.

1. The "test_output" directory is large, and isn't something that needs to be installed by users. The files in this directory could be removed to make it smaller and reducing installation time.
2. The .fits files in the "tests" directory are also quite large and could be reduced to smaller files.
3. There is a duplicate directory named "LATTE_imgs" in tests

## Functionality

I started running `LATTE` and went through the initial downloading set up. It wasn't very clear on what was being downloaded at this point (I believe it's essentially the manifest of target pixel files and light curve files from MAST). As I was waiting, I was a bit worried it was going to try and download all the light curves and target pixel files. It might be beneficial to add some additional documentation on what `LATTE` is downloading and why.

When I ran a test target and closed the first resulting images, there was a long pause while the `LATTE` downloaded other products. I believe it downloaded the TPF. It would be useful to have somewhere in the documentation the explicit workflow for `LATTE`, i.e. what is being downloaded, why, and when.

Some more in depth comments about the functionality

1. I am not familiar with what the "Aperture Size" test is in this context, can you add in your report or documentation what it is designed to show?

2. I am not clear why there is a GUI for the plot window, as the point of `LATTE` seems to be to create a data validation report for TESS data. It seems like these plots should be generated and closed, rather than left up to the user to close? I see in the GUI I think this can be turned off, perhaps it should be turned off by default?

## Documentation

It is not clear from the documentation what the express purpose of `LATTE` is. I do see there is a "blurb" in the README, but that didn't make clear to me what `LATTE` was doing, (i.e. downloading data products, producing certain diagnostic plots, and producing a report). It would be beneficial for the user to add a more in-depth explanation.

I find the documentation for the GUI could be improved, could you add to the READE me with a diagram highlighting each section of the GUI and what it will do?

I feel the user may benefit from having an example workflow for using `LATTE` in the documentation.

As I've said above, some documentation on what is being downloaded from MAST, why, typical sizes, typical times etc would improve the documentation for new users.

The documentation on the output plots either in the README or the report could be improved to explain what each plot is designed to show the user.
