### 2020-03-09 #2101 @christinahedges
Thanks @arfon.

I've taken a look at LATTE and added my review comments to an issue [here](https://github.com/noraeisner/LATTE/issues/3). @noraeisner and @xuanxu, I have left comments there why some of the boxes are unticked.

I find the biggest comment I have to be the state of the field in the paper needs to be improved. I think that the SPOC validation reports, MIT QLP validation reports, eleanor, and lightkurve are all relevant context that should be mentioned. I also find in general that the documentation could be improved a little to show the workflow of an average user, and discuss what each validation plot is designed to show.
