### 2020-03-16 #2101 @aureliocarnero
Hello @noraeisner, I have no capabilities at the moment to run the mpi example.... Nonetheless, we can pause now the review. Don't worry about this now. Take care and speak soon!
