### 2020-03-16 #2101 @noraeisner
Hi all, thank you so much for your review comments and I'm so sorry for my delayed reply. It took me a couple of days to relocate (I'm in isolation in Belgium now) but I will be working on the comments today and I'll respond to the individual comments as I address them. Of course I also completely understand if the reviewers prefer to pause the process for now. I hope everyone is safe and well!
