### 2020-03-16 #2101 @aureliocarnero
Im glad you are good! We in Spain are isolated as well now... So I'm happy to continue with the validation. Fortunately I'm healthy and I haven't been affected at all so far.
Apart from Christina's comment that I agree with, I just need to look at the parallel example now. I'm happy to continue with the validation until the end....
Let me see if I can finish, my computer at home is not the best though...
