### 2020-03-16 #2101 @aureliocarnero
Hello @noraeisner so I tested the script example_parallelized_code.py and even though I have MPI installed, I do not have more than 1 core, therefore it enters into the serial part of the code. Here, I have this error. Could you please help me?

Traceback (most recent call last):
  File "example_parallelized_code.py", line 250, in <module>
    transit_list = ast.literal_eval(((df.loc[df['TIC_ID'] == f]['db_peak']).values)[0])
  File "/home/carnero/anaconda3/envs/mypython3/lib/python3.8/site-packages/pandas/core/frame.py", line 2800, in __getitem__
    indexer = self.columns.get_loc(key)
  File "/home/carnero/anaconda3/envs/mypython3/lib/python3.8/site-packages/pandas/core/indexes/base.py", line 2648, in get_loc
    return self._engine.get_loc(self._maybe_cast_indexer(key))
  File "pandas/_libs/index.pyx", line 111, in pandas._libs.index.IndexEngine.get_loc
  File "pandas/_libs/index.pyx", line 138, in pandas._libs.index.IndexEngine.get_loc
  File "pandas/_libs/hashtable_class_helper.pxi", line 1618, in pandas._libs.hashtable.PyObjectHashTable.get_item
  File "pandas/_libs/hashtable_class_helper.pxi", line 1626, in pandas._libs.hashtable.PyObjectHashTable.get_item
KeyError: 'db_peak'
