+++ paper/paper.bib
@@ -452,3 +452,45 @@ archivePrefix = {arXiv},
   year={2010},
   publisher={American Association for the Advancement of Science}
 }
+
+@INPROCEEDINGS{2019QLP,
+       author = {{Huang}, Xu and {Burt}, Jennifer and {Vanderburg}, Andrew and
+         {Gunther}, Maximillian and {Shporer}, Avi and {Dittmann}, Jason and
+         {Winn}, Joshua},
+        title = "{A Quick look into the first discoveries of TESS}",
+    booktitle = {American Astronomical Society Meeting Abstracts \#233},
+         year = 2019,
+       series = {American Astronomical Society Meeting Abstracts},
+       volume = {233},
+        month = jan,
+          eid = {209.08},
+        pages = {209.08},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019AAS...23320908H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+@ARTICLE{2019eleanor,
+       author = {{Feinstein}, Adina D. and {Montet}, Benjamin T. and
+         {Foreman-Mackey}, Daniel and {Bedell}, Megan E. and
+         {Saunders}, Nicholas and {Bean}, Jacob L. and
+         {Christiansen}, Jessie L. and {Hedges}, Christina and {Luger}, Rodrigo and
+         {Scolnic}, Daniel and {Cardoso}, Jos{\'e} Vin{\'\i}cius de Miranda},
+        title = "{eleanor: An Open-source Tool for Extracting Light Curves from the TESS Full-frame Images}",
+      journal = {\pasp},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2019,
+        month = sep,
+       volume = {131},
+       number = {1003},
+        pages = {094502},
+          doi = {10.1088/1538-3873/ab291c},
+archivePrefix = {arXiv},
+       eprint = {1903.09152},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019PASP..131i4502F},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
