+++ paper/paper.md
@@ -35,8 +35,9 @@ NASA's Transiting Exoplanet Survey Satellite [*TESS*, @ricker15] is the first ne
 
 Unlike its predecessor *Kepler* [@borucki2010kepler], *TESS* uses very large pixels, meaning that astrophysical false positives (e.g. where the signal of a target is contaminated by a faint eclipsing binary) are expected to be common. Many false positive scenarios can be ruled out from detailed examination of the *TESS* data using standard diagnostic tests. 
 
-``LATTE`` is an open source Python package that performs these standard diagnostic tests to help identify, vet and characterise signals in the *TESS* lightcurves in order to weed out instrumental and astrophysical false positives. The program is designed to allow for a fast, in depth analysis of targets that have already been identified as promising candidates by the main *TESS* pipelines or via alternative methods such as citizen science [e.g., @fischer12; @Christiansen2018; @eisner19]. The implemented diagnostic tests are similar to the ‘Data Validation’ steps performed by the Science Processing Operations Center [SPOC; @jenkins16] pipeline. However, while the SPOC pipeline tests are performed on a selection of two-minute cadence *TESS* targets only, ``LATTE`` allows for the analysis of any *TESS* target with a valid TIC ID [@Stassun19], including both the two-minute and thirty-minute cadence targets.
+``LATTE`` is an open source Python package that performs these standard diagnostic tests to help identify, vet and characterise signals in the *TESS* lightcurves in order to weed out instrumental and astrophysical false positives. The program is designed to allow for a fast, in depth analysis of targets that have already been identified as promising candidates by the main *TESS* pipelines or via alternative methods such as citizen science [e.g., @fischer12; @Christiansen2018; @eisner19]. The implemented diagnostic tests are similar to the ‘Data Validation’ steps performed by the Science Processing Operations Center [SPOC; @jenkins16] pipeline for the short-cadence data, and by the quick-look pipeline (QLP; @2019QLP) for the long-cadence data. However, while the SPOC pipeline tests are performed on a selection of two-minute cadence *TESS* targets only, ``LATTE`` allows for the analysis of any *TESS* target with a valid TIC ID [@Stassun19], including both the two-minute and thirty-minute cadence targets.
 
+Other tools that are often used to extract and vet TESS light curves include the open source packeges *Lightkurve* [@2018lightkurve] and *eleanor* [@2019eleanor], for the long and short-cadence data respectively. These tools can perform similar diagnostic tests to the ones provided by ``LATTE``, however, ``LATTE`` differs in that it does not require any interaction with python coding, and summarizes multiple diagnostic tests in an automated manner. 
 
 # ``LATTE`` 
 
@@ -44,12 +45,11 @@ The standard vetting tests performed by ``LATTE`` provide a useful tool for both
 
 Upon execution, ``LATTE`` prompts the user enter the TIC ID [@Stassun19] of a target star, as well as the observational sector(s) to analyse. The corresponding data are then downloaded directly from the Mikulski Archive for Space Telescopes (MAST, http://archive.stsci.edu/tess/) and displayed in a Graphical User Interface (GUI), as shown in Figure 1. The user is asked to identify the times of the transit-like events by directly clicking on the lightcurve plots and recording the transit-times using the 'Add time' button. Additional settings, such as whether to run a Box-Least-Squares (BLS) algorithm that searches for periodic signals in the lightcurve, can be selected within the GUI (see left side of Figure 1). Once the times of one or more transit-like events have been identified the program automatically downloads and handles the *TESS* data to provide standard exoplanet diagnostic plots and statistics.
 
[…]
