+++ LATTE/LATTEutils.py
@@ -38,8 +38,41 @@ from matplotlib.widgets import Slider, Button, RadioButtons, TextBox, CheckButto
 from LATTE import filters
 from LATTE import LATTEbrew as brew
 
+'''
+Overview of LATTE scipts:
+
+__main__.py      : Intitialises the parameters, what TIC ID, sector, checks for downloads of the data, FFI or not? 
+
+LATTEutils.py    : All the functions needed to download data and text files, runs the interactive gui, all of the plotting and data handling. 
+
+LATTE_DV.py      : Scipt to combine all of the results from LATTEbrew in order to generate a pdf data validation report.
+
+LATTEbrew.py     : Calls the LATTEutils functions in turn in order to store the results and keeps track of what has been generated. Calls the LATTE_DV.py function to collate all the results.
+
+
+The functions in this scipt are organised as follows: 
+
+1) 3 functions that call the interative GUI plotting for the short cadence and FFI data. There are two functions for the FFI, one that asks the user to 
+identify the aperture sizes manually and one that does it automatically. These functions are called from __main__.py and call the 'brew' scipt at the end. 
+The brew scipt then continues with collating the data by calling the remaining functions in this script. 
+
+2) 4 functions that download the text files needed for the LATTE to run. These are only exectuted the first time that LATTE is run, and when new TESS data is made available. 
+These functions are called from within __main__.py
+
+3) 3 functions that determine the sectors in which the target was observed and TICs of nearest neighbours. 
+
+4) 8 functions needed to download and process the data, both from the FFIs and short cadence data. 
+These functions are called from within LATTEbrew.py and from withint the interact functions in this script.
+
+5) 13 functions for plotting the results, both from the FFIs and short cadence data. 
+These functions are called from within LATTEbrew.py.
+
+4) 4 short functions that are needed in various other functions within this scipt, such as one to bin the data. 
+
+'''
+
 # check whether pyaneti has been sucessfully installed - if it has not been installed, don't give the option to model the data
-# NOTE: in this version the Pyaneti modeling will not work (except on my computer) - this is being handled in the next release.
+# NOTE: in this version the Pyaneti modeling will not work (except on Nora's computer) - this is being handled in the next release.
 
 try:
     from LATTE import pyaneti_LATTE
@@ -52,6 +85,7 @@ except:
 # --------------------------------------------
 
 # the main interactive tool used to identify the times of the transit-like events
+# This function is called from __main__.py
 def interact_LATTE(tic, indir, syspath, sectors_all, sectors, ra, dec, args):
     
     '''
@@ -82,7 +116,7 @@ def interact_LATTE(tic, indir, syspath, sectors_all, sectors, ra, dec, args):
     # ---------------   
 
     # call function to download the data from MAST
-    print ("Start data download.....", end =" ")
+    print ("Start lightcurve data download.....", end =" ")
     alltime, allflux, allflux_err, all_md, alltimebinned, allfluxbinned, allx1, allx2, ally1, ally2, alltime12, allfbkg, start_sec, end_sec, in_sec, tessmag, teff, srad = download_data(indir, sectors, tic)
     print ("done.\n")
     
@@ -272,11 +306,11 @@ def interact_LATTE(tic, indir, syspath, sectors_all, sectors, ra, dec, args):
     # only give the model option if pyaneti has been sucessfully installed
     if pyaneti_installed == True:
         var_ax = fig.add_axes([0.025, 0.34, 0.119, 0.21]) # x, y, width, height
[…]
