+++ MANIFEST.in
@@ -1,3 +1,1 @@
 graft LATTE/LATTE_imgs 
-graft LATTE/tests 
-graft LATTE/test_output

[…]
