### 2020-03-17 6aa4aaa
sector selection option
