+++ LATTE/LATTEutils.py
@@ -27,6 +27,10 @@ from astropy.stats import BoxLeastSquares
 from matplotlib.patches import Rectangle
 from lightkurve import TessTargetPixelFile
 
+import matplotlib.widgets
+import matplotlib.patches
+import mpl_toolkits.axes_grid1
+
 from tess_stars2px import tess_stars2px_function_entry
 from reproject import reproject_interp, reproject_exact
 from reproject.mosaicking import find_optimal_celestial_wcs
@@ -37,6 +41,8 @@ from matplotlib.widgets import Slider, Button, RadioButtons, TextBox, CheckButto
 # custom modules
 from LATTE import filters
 from LATTE import LATTEbrew as brew
+
+
 
 '''
 Overview of LATTE scipts:
@@ -80,6 +86,7 @@ try:
 except:
     pyaneti_installed = False
 
+pyaneti_installed = False
 # --------------------------------------------
 # -----------------interact ------------------
 # --------------------------------------------
@@ -114,6 +121,8 @@ def interact_LATTE(tic, indir, syspath, sectors_all, sectors, ra, dec, args):
             new_shape[1], arr.shape[1] // new_shape[1])
         return arr.reshape(shape).mean(-1).mean(1)
     # ---------------   
+    # this needs to be global for the sector slider
+    global in_sec
 
     # call function to download the data from MAST
     print ("Start lightcurve data download.....", end =" ")
@@ -128,7 +137,7 @@ def interact_LATTE(tic, indir, syspath, sectors_all, sectors, ra, dec, args):
     # Plot the interactive plot - uses matplolib
     # -------------------------
     
-    fig, ax = plt.subplots(2, 1, figsize=(10,7))
+    fig, ax = plt.subplots(2, 1, figsize=(11,7.5))
     plt.tight_layout()
 
     # Adjust the plots region to leave some space for the sliders and buttons
@@ -163,6 +172,87 @@ def interact_LATTE(tic, indir, syspath, sectors_all, sectors, ra, dec, args):
         return [time_binned, flux_binned]
     # ---------------
 
+    # define the slider that let's you chose the sector to look at 
+    class PageSlider(matplotlib.widgets.Slider):
+    
+        def __init__(self, ax, label, numpages = 10, valinit=0, valfmt='%1d', 
+                     closedmin=True, closedmax=True,  
+                     dragging=True, **kwargs):
+        
+            self.facecolor=kwargs.get('facecolor',"w")
+            self.activecolor = kwargs.pop('activecolor',"b")
+            self.fontsize = kwargs.pop('fontsize', 10)
+            self.numpages = numpages
+    
+            super(PageSlider, self).__init__(ax, label, 0, numpages, 
+                                valinit=valinit, valfmt=valfmt, **kwargs)
+        
+            # make it so that numpages is the index of the array of the 
+            global in_sec
+
+            self.poly.set_visible(False)
+            self.vline.set_visible(False)
+            self.pageRects = []
+            for i in range(numpages):
+
+                facecolor = self.activecolor if i==valinit else self.facecolor
+                r  = matplotlib.patches.Rectangle((float(i)/numpages, 0), 1./numpages, 1, 
+                                    transform=ax.transAxes, facecolor=facecolor)
+                ax.add_artist(r)
+                self.pageRects.append(r)
+
+                seclist_names = ['All'] + in_sec
+
+                sec_i = seclist_names[i]
+                
+                ax.text(float(i)/numpages+0.5/numpages, 0.5, str(sec_i),  
+                        ha="center", va="center", transform=ax.transAxes,
+                        fontsize=self.fontsize)
+            self.valtext.set_visible(False)
+    
[…]
