+++ README.md
@@ -59,7 +59,7 @@ In *FFI mode* the data is downloaded using TESScut and the data detrended using 
 
 ## Input target list
 
-In order to efficiently generate diagnostic plots for multiple targets without having to interactively enter the target TIC IDs, ``LATTE`` can be executed with an input file that lists the TIC IDs, the times of the transit-like events, and the observational sectors to consider (if known). See example. In the cases where the times of the transit events or the sectors have not been entered, the user is prompted to enter these manually via the GUI, as descibed above. For longer target lists the code can also be parallelized (see example). If you want to run it with this mode, enter: 
+In order to efficiently generate diagnostic plots for multiple targets without having to interactively enter the target TIC IDs, LATTE can be executed with an input file that lists the TIC IDs, the times of the transit-like events, and the observational sectors to consider (if known). See example. In the cases where the times of the transit events or the sectors have not been entered, the user is prompted to enter these manually via the GUI, as descibed above. For longer target lists the code can also be parallelized (see example). If you want to run it with this mode, enter: 
 
 	python3 -m LATTE --targetlist=path/to/the/csv/input/file
 
@@ -70,7 +70,6 @@ The program will skip targets that have already been analysed by you, so if you 
 Once you have identified the TIC ID, the observational sectors and the aperture sizes (*FFI mode* only), you will see a screen that has the full lighcurve as well as a zoom in of the lightcurve (see figure below). The solid red line across the full lightcurve lets you know where on the lighcurve you are zooming in on. Click on the top (full) or bottom (zoomed in) plots to change the location of the zoom in until the red vertical line is centred on the mid-point of a transit-event. When you are happy with the location press the 'Add Time' button below the plots in order to record this time (in TBJD). You can delete wrongly entered times with the 'Delete time'. The saved times will be shown on the screen. The position of the red line can also be changed by dragging the teal coloured 'Transit' slider with your mouse. The y-scale of the plots can be changed with the grey coloured slider.
 
 ![Interface Example](https://github.com/noraeisner/LATTE/blob/master/paper/interface_explained.png)
-
 
 Additional options are displayed to the left of the plots.
 
@@ -97,33 +96,34 @@ The code will then generate download and process all of the data. Note that all 
 
 ### Output
 
+For more information on how to interpret the results shown in the reports look at the 'How to make a LATTE' document in the 'example' folder. 
 
 **Figures:**
 
 - Full lightcurve with the times of the momentum dumps marked. 
 
-![Full LC](https://github.com/noraeisner/LATTE/blob/master/example_output/94986319_fullLC_md.png)
+![Full LC](https://github.com/noraeisner/LATTE/blob/master/example/example_output/94986319_fullLC_md.png)
 
 The solid red lines at the bottom of the figure indicated the times of the reaction wheel momentum dumps and the dashed black line(s) show the times of the marked transit events. Momentum dumps occur around every 2 to 2.5 days and typically last around half an hour. This can affect the satellite pointing, resulting in spurious transit-like signals. 
 	
 
 - Background flux around the times of the marked transit event(s).
 
-![Background Flux](https://github.com/noraeisner/LATTE/blob/master/example_output/94986319_background.png)
+![Background Flux](https://github.com/noraeisner/LATTE/blob/master/example/example_output/94986319_background.png)
 
 Enhanced scattered light in the telescope optics can cause the background flux to rise sharply each time TESS approaches the perigee of its eccentric orbit around the Earth. Additionally, solar system objects, such as asteroids passing through the field of view will show up as a spike in the background flux. Both of these can result in a transit-like signal in the light curve. 
 
 
[…]
