### 2020-03-18 noraeisner/LATTE#3 @noraeisner
Hi @christinahedges,

Thank you so much for your thorough review. I address the points one by one below:

Installation
1) test_output directory. I have removed the test_output file from the pip install download. It remains on the github page so that anyone who wants to test the code can download that version. The LATTE_imgs folder is repeated in the test file so that all of the test files are self contained in one folder.

Functionality
2) Data Download. That is a great point, thank you, I definitely don't want anyone to think that all the data is being downloaded. I have made this clearer on the readme page and changed the text that is printed to the command line to make it clearer that no actual TESS data is being downloaded.

3) Data Download. I have added a section to the readme file called 'LATTE workflow - what is being downloaded, when and why?' that explains the workflow and outlines when and why data is being downloaded and used and how big the files typically are (I'm hesitant to give times as they vary drastically depending on the internet speed). I also added a very short section that explains the functionalities of the four main LATTE scripts to the readme and more detailed descriptions at the start of the main .py files. Please do let me know if you had something else in mind to make it more clear.

4) Aperture Size test. I now explain all of the diagnostic plots in more detail in the readme.  I have also created a document that explains how to interpret the plots in more detail with examples of 'good' and 'bad' planet candidates - this document is in 'examples' folder. I have also added short video that shows how to run LATTE.

5) GUI. I completely agree with you that the point is to not interact with the code. The default is now such that the plots are *not* displayed and changed the option in the GUI such that the user can choose to display by selecting "Show Plots".

Documentation
6) A statement of need. I have expanded on the blurb in the readme file to make the prupose of the code more clear.

7) Example usage. There is now a video of how LATTE is used, example output files and a document of how to interpret the results in order to identify planet candidates (in the 'example' folder).

8) Functionality documentation. Please see point 3).

9) Functionality documentation. The plots and downloads were definitely lacking explanation, I'm so sorry. I have written a short summary of each plot in the readme along with the example images. There are also more examples and more explanation of each plot in the 'example' folder.

10) Automated tests. I'm not sure why these aren't working but I'm looking into them today, sorry.

11) Other comments. I have added an annotated image of the GUI interface explaining the different functionalities.

Software paper
12) State of the field. I have added a paragraph to the paper mentioning the other tools that are available and how LATTE complements and differs from them. I haven't added very much text, because JOSS paper guidelines ask for a paper length between 250-1000 words, which I am already significantly over, but obviously do let me know if you would like the introduction to be more in depth and I'd be more than happy to add more.
