### 2020-03-23 590da5a
parallelized code
