+++ example_parallelized_code.py
@@ -28,7 +28,6 @@ import sys
 # get the system path
 syspath = str(os.path.abspath(LATTEutils.__file__))[0:-14]
 
-
 # set up for supercomputer
 try:
 	from mpi4py import MPI
@@ -54,7 +53,21 @@ ttran = 0.1
 
 # -------------------
 
-def process(indir, sectors, tic, transit_list, args):
+def process(indir, tic, sectors_in, transit_list, args):
+
+	sectors_all, ra, dec = LATTEutils.tess_point(indir, tic) 
+
+	try:
+		# Sucessfully entered sectors
+		# check that the target was actually observed in the stated sector
+		sectors = list(set(sectors_in) & set(sectors_all))
+
+		if len(sectors) == 0:
+			print ("The target was not observed in the sector(s) you stated ({}). \
+					Therefore take all sectors that it was observed in: {}".format(sectors, sectors_all))
+			sectors = sectors_all
+	except:
+		sectors = sectors_all
 
 	# download the data 
 	alltime, allflux, allflux_err, all_md, alltimebinned, allfluxbinned, allx1, allx2, ally1, ally2, alltime12, allfbkg, start_sec, end_sec, in_sec, tessmag, teff, srad = LATTEutils.download_data(indir, sectors, tic)
@@ -69,7 +82,7 @@ def process(indir, sectors, tic, transit_list, args):
 	save = True
 	DV = True
 
-	sectors_all, ra, dec = LATTEutils.tess_point(indir, tic)
+	
 
 	# generate the plots
 	try:
@@ -173,12 +186,12 @@ def process(indir, sectors, tic, transit_list, args):
 
 if __name__ == '__main__':
 	ap = ArgumentParser(description='Script to pLot TESS LCs for Zooniverse project')
-	ap.add_argument('--sector', type=int, help='The sector to analyse', default=None)
+	#ap.add_argument('--sector', type=int, help='The sector to analyse', default=None)
 	ap.add_argument('--number', type=int, help='The number of randomly selected subjects (default all in indir folder)', default=None)
+	ap.add_argument('--targetlist', type=str, help='the link to the target file list', default='no')
 
 	# these just have to be here
 	#ap.add_argument('--sector', type=str, help='the sector(s) to look at', default = 'no')
-	ap.add_argument('--targetlist', type=str, help='the link to the target file list', default='no')
 	ap.add_argument('--noshow', action='store_true', help='if you want to NOT show the plots write --noshow in the command line')
 	ap.add_argument('--o', action='store_true', help='if you call this old files will be overwriten in the non-interactive version')
 	ap.add_argument('--auto', action='store_true', help='automatic aperture selection')
@@ -191,22 +204,82 @@ if __name__ == '__main__':
 
 	args = ap.parse_args()
 
-	sector = args.sector
-	sectors = [args.sector]
+	#sector = args.sector
+	#sectors = [args.sector]
 
 	# make sure that they don't show...
 	args.noshow = True
 	args.mpi = True
 
-	indir = 'path/to/input/file'
+	# check whether a path already exists
+	if not os.path.exists("{}/_config.txt".format(syspath)):
+	
+		# if it doesn't exist ask the user to put it in the command line
+		indir = input("\n \n No output path has been set yet. \n \n Please enter a path to save the files (e.g. ./LATTE_output or /Users/yourname/Desktop/LATTE_output) : " )
+	
+		# SAVE the new output path
+		with open("{}/_config.txt".format(syspath),'w') as f:
+			f.write(str(indir))
+		
+		print("\n New path: " + indir)
+	
[…]

[…]
