### 2020-03-23 #2101 @noraeisner
Hello @aureliocarnero, I'm glad to hear that you are safe and well in Spain! Belgium has now also entered a lock-down which is definitely a strange experience.

I'm really sorry, the mpi code that was on github before required the user to enter their own file structure within the code itself. I have changed it to make it much more general and it now works with the example input file - I have added a short section to the readme file of how to execute it.

I also made a small change to the GUI. If more than one sector is now identified, you can chose to look at the sectors one-by-one or all of them together.  I made this change as before it was quite difficult to see what was going on if too many sectors were selected due to the figure dimensions.  Nothing in the code itself has changed except the way that the interface looks.

I have also expanded on the instructions in the readme and added some more background information to the paper (as outlined here https://github.com/noraeisner/LATTE/issues/3).
