+++ READMEpip.md
@@ -4,9 +4,6 @@ Written by Nora L. Eisner
 
 email: *nora.eisner@new.ox.ac.uk*
 
-![LATTE logo](https://github.com/noraeisner/LATTE/blob/master/LATTE/LATTE_imgs/LATTE_logo_small.png)
-
-
 ### THE CODE
 
 --------
@@ -24,7 +21,7 @@ The code is designed to be fully interactive and does not require any prior know
 
 ### Installation
 
-LATTE requires python3 to be installed on your computer, which can be download from https://www.python.org/downloads/. Once you have python3, you can download the code directly from github. Alternatively you can install LATTE using pip (https://pypi.org/project/tessLATTE/) via your command line with:
+LATTE requires python3 to be installed on your computer, which can be download from https://www.python.org/downloads/. Once you have python3, you can download the code directly from the LATTE github page. Alternatively you can install LATTE using pip (https://pypi.org/project/tessLATTE/) via your command line with:
 
 	pip3 install tessLATTE      
 
@@ -55,7 +52,7 @@ The '*normal mode*' looks at the short-cadence *TESS* data which has already bee
 
 In *FFI mode* the data is downloaded using TESScut and the data detrended using PCA, a moving average filter and k-sigma clipping. Unlike the *TESS* 2-minute cadence targets, the FFIs do not come with a pre-chosen optimal aperture. By default, the user is given the option to manually select the pixels within each aperture for a 'large' and a 'small' aperture. This GUI opens automatically and the two apertures are selected by clicking on the desired pixels in the interface (see image below). The two (large and small aperture) lightcurves are simultaneously displayed. When the FFI mode is run with the command '--auto', the aperture size is chosen manually using a threshold flux value centred at the midpoint of the postage stamp.
 
-![FFI Example](https://github.com/noraeisner/LATTE/blob/master/paper/Fig2.png)
+Please see the LATTE github page for an example of what the interface looks like.
 
 ## Input target list
 
@@ -69,9 +66,9 @@ The program will skip targets that have already been analysed by you, so if you 
 
 Once you have identified the TIC ID, the observational sectors and the aperture sizes (*FFI mode* only), you will see a screen that has the full lighcurve as well as a zoom in of the lightcurve (see figure below). The solid red line across the full lightcurve lets you know where on the lighcurve you are zooming in on. Click on the top (full) or bottom (zoomed in) plots to change the location of the zoom in until the red vertical line is centred on the mid-point of a transit-event. When you are happy with the location press the 'Add Time' button below the plots in order to record this time (in TBJD). You can delete wrongly entered times with the 'Delete time'. The saved times will be shown on the screen. The position of the red line can also be changed by dragging the teal coloured 'Transit' slider with your mouse. The y-scale of the plots can be changed with the grey coloured slider.
 
-![Interface Example](https://github.com/noraeisner/LATTE/blob/master/paper/interface_explained.png)
+Please see the LATTE github page for an example of what the interface looks like.
 
-Additional options are displayed to the left of the plots.
+Additional options are displayed to the left of the plot.
 
 Sector Selection: 
 - The sectors that you have chosen to look at are listed in the 'Sector' bar below the plot. You can change the sector to look at by clicking on the corresponding number in the bar, or on 'all' too look at all of them. Alternatively you can use the arrows to the right of the bar to scroll through them. Only the top panel will change. 
@@ -96,64 +93,46 @@ The code will then generate download and process all of the data. Note that all 
 
 ### Output
 
-For more information on how to interpret the results shown in the reports look at the 'How to make a LATTE' document in the 'example' folder. 
+For more information on how to interpret the results shown in the reports look at the 'How to make a LATTE' document in the 'example' folder. Please see the LATTE github page for example plots.
 
 **Figures:**
 
[…]
