### 2020-03-23 #2101 @aureliocarnero
@noraeisner great improvement! Im my case, we are very near the end of the validation. I'm not English speaker, so in my opinion the paper is fine, but please consider the other referee questions as well.
Just a final comment, I found a bug in the example_parallelized_code.py code.
Where it reads:
                utils.data_files(indir)
                utils.tp_files(indir)
                utils.TOI_TCE_files(indir)
                utils.momentum_dumps_info(indir)

I think it should read:

          LATTEutils.data_files(indir)
          LATTEutils.tp_files(indir)
          LATTEutils.TOI_TCE_files(indir)
          LATTEutils.momentum_dumps_info(indir)

Correct? Could you check that for me? And let me know when you commit the change for a final test?
