+++ tessLATTE.egg-info/SOURCES.txt
@@ -19,73 +19,6 @@ LATTE/LATTE_imgs/LATTE_logo_small.png
 LATTE/LATTE_imgs/PHT_logo.jpg
 LATTE/LATTE_imgs/TESS_logo.png
 LATTE/pyaneti_LATTE/__init__.py
-LATTE/test_output/.DS_Store
-LATTE/test_output/55525572/.DS_Store
-LATTE/test_output/55525572/55525572_aperture_size.png
-LATTE/test_output/55525572/55525572_apertures_0.png
-LATTE/test_output/55525572/55525572_background.png
-LATTE/test_output/55525572/55525572_bls_first.png
-LATTE/test_output/55525572/55525572_bls_second.png
-LATTE/test_output/55525572/55525572_centroids.png
-LATTE/test_output/55525572/55525572_data.txt
-LATTE/test_output/55525572/55525572_data_pyaneti.dat
-LATTE/test_output/55525572/55525572_fit_test.png
-LATTE/test_output/55525572/55525572_flux_comparison.png
-LATTE/test_output/55525572/55525572_fullLC_md.png
-LATTE/test_output/55525572/55525572_individual_pixel_LCs_0.png
-LATTE/test_output/55525572/55525572_individual_pixel_LCs_1.png
-LATTE/test_output/55525572/55525572_individual_pixel_LCs_16380.png
-LATTE/test_output/55525572/55525572_nearest_neighbours.png
-LATTE/test_output/55525572/55525572_phase_folded.png
-LATTE/test_output/55525572/55525572_star_field.png
-LATTE/test_output/55525572/DV_report_55525572.pdf
-LATTE/test_output/55525572/individual_pixel_LCs.png
-LATTE/test_output/55525572/model_out/55525572-trdata.txt
-LATTE/test_output/55525572/model_out/55525572-trmodel.txt
-LATTE/test_output/55525572/model_out/55525572_all_data.dat
-LATTE/test_output/55525572/model_out/55525572_init.dat
-LATTE/test_output/55525572/model_out/55525572_parameters.csv
-LATTE/test_output/55525572/model_out/55525572_params.dat
-LATTE/test_output/55525572/model_out/55525572_params.tex
-LATTE/test_output/55525572/model_out/55525572_posterior.pdf
-LATTE/test_output/55525572/model_out/55525572_posterior.png
-LATTE/test_output/55525572/model_out/55525572b_tr.pdf
-LATTE/test_output/55525572/model_out/55525572b_tr.png
-LATTE/test_output/data/TOI_list.txt
-LATTE/test_output/data/tesscurl_sector_all_dv.sh
-LATTE/tests/.DS_Store
-LATTE/tests/__init__.py
-LATTE/tests/test_BLS.py
-LATTE/tests/test_DVreport.py
-LATTE/tests/test_data_download.py
-LATTE/tests/test_data_products.py
-LATTE/tests/test_plotting.py
-LATTE/tests/test_server_response.py
-LATTE/tests/tic55525572_lc.fits
-LATTE/tests/tic55525572_tp.fits
-LATTE/tests/LATTE_imgs/LATTE_logo.png
-LATTE/tests/LATTE_imgs/PHT_logo.jpg
-LATTE/tests/LATTE_imgs/TESS_logo.png
-LATTE/tests/__pycache__/__init__.cpython-35.pyc
-LATTE/tests/__pycache__/__init__.cpython-37.pyc
-LATTE/tests/__pycache__/test_BLS.cpython-35.pyc
-LATTE/tests/__pycache__/test_BLS.cpython-37.pyc
-LATTE/tests/__pycache__/test_DVreport.cpython-35.pyc
[…]
