+++ LATTE/_config.txt
@@ -1,0 +1,1 @@
+/Users/Nora/Documents/research/TESS/planethunters/code/LATTE/LATTE/LATTE_output
