+++ MANIFEST.in
@@ -1,1 +1,1 @@
-graft LATTE/LATTE_imgs 
+graft LATTE/LATTE_imgs
