+++ README.md
@@ -1,6 +1,5 @@
-# LATTE
+# LATTE  -  Lightcurve Analysis Tool for Transiting Exoplanet
 
-## Lightcurve Analysis Tool for Transiting Exoplanet
 
 Written by Nora L. Eisner
 
