+++ LATTE/LATTEutils.py
@@ -2846,26 +2846,26 @@ def download_data_FFI_interact(indir,sector, sectors_all, tic, save = False):
             print ("ERROR: Could not download the tpf file at this time. Please try again.")
             sys.exit('')
         # ---------
-        
+
         # get rid of the 'bad' quality data - use the data flags and only take data where the quality = 0. 
         try:
             quality = tpf.quality
             tpf = tpf[quality == 0]
         except:
             print ('This file has no quality flag.')
-            
+
         # save the raw tpf (target pixel file) data which will be used later - need to store it in this format as it will be used for the reprojection into the right coordinates.
         tpf_list.append(tpf)
-        
+
         print ("done.\n")
-        
+
         # extract the information and perform PCA
         print ("Start PCA analysis...", end =" ")
         
         # acess the flux information
         X1 = tpf.flux
         X1flux_list.append(X1)  # store tghe flux array in the list
-        
+
         arrshape_list.append(X1.shape)  # store the shape in the list - not always the same shape but mostly 11 x 11 pixels.
         
         # calculate the average pixel value across all frames. 
@@ -3817,7 +3817,7 @@ def data_bls(tic, indir, alltime, allflux, allfluxbinned, alltimebinned, args):
         normalized binned flux for the LC of the target
 
     Returns
-    -------    
+    -------
         two lists of the statistics of the to BLS runs. Each list contains:
     stats_period
     stats_t0
@@ -3885,7 +3885,7 @@ def data_bls(tic, indir, alltime, allflux, allfluxbinned, alltimebinned, args):
     alltimebinned = np.array(alltimebinned)[mask_binned]
     allfluxbinned = np.array(allfluxbinned)[mask_binned]
     # -----------------------
-
+    
     durations = np.linspace(0.05, 0.5, 15) # ????? CHECK THESE 
     periods = np.arange(0.51, (np.nanmax(alltimebinned) - np.nanmin(alltimebinned)), 0.01)
 
