+++ setup.py
@@ -5,12 +5,12 @@ from setuptools import setup, find_packages
 HERE = pathlib.Path(__file__).parent
 
 # The text of the README file
-README = (HERE / "README.md").read_text()
+README = (HERE / "READMEpip.md").read_text()
 
 # This call to setup() does all the work
 setup(
     name="tessLATTE",
-    version="0.1.16",
+    version="0.1.17",
     description="Lightcurve Analysis Tool for Transiting Exoplanets",
     long_description=README,
     long_description_content_type="text/markdown",
