+++ lasfk/data/tesscurl_sector_all_tp.sh
@@ -1,0 +1,1 @@
+#all LC file links
