### 2020-04-20 #2101 @noraeisner
@aureliocarnero, thank you for seeing that bug, I have changed those lines in example_parallelized_code.py. The final version has been uploaded to Github and the pip version updated (now version 0.1.18) for whenever you have the time. I have addressed Christina comments and am waiting to hear back. Thank you so much.
