### 2020-05-12 #2101 @aureliocarnero
Thank you for the reminder. I will check ASAP.

El mar., 12 may. 2020 a las 10:13, Juanjo Bazán (<notifications@github.com>)
escribió:

> Hi @christinahedges <https://github.com/christinahedges> @aureliocarnero
> <https://github.com/aureliocarnero>, hope you are safe and well. Did you
> have the time to check the latest changes Nora made?
>
