### 2020-05-12 #2101 @aureliocarnero
So @noraeisner I found another error when running both the parallelized example and the unittests. There is an error with download and database issues...

ERROR: test_downloadLC (LATTE.tests.test_data_download.TestDownloadLC)
Traceback (most recent call last):
  File "/home/carnero/codes/LATTE/LATTE/tests/test_data_download.py", line 71, in test_downloadLC
    alltime, allflux, allflux_err, all_md, alltimebinned, allfluxbinned, allx1, allx2, ally1, ally2, alltimel2, allfbkg, start_sec, end_sec, in_sec, tessmag, teff, srad = LATTEutils.download_data(outdir, [5], '55525572', binfac = 5)
  File "/home/carnero/codes/LATTE/LATTE/LATTEutils.py", line 2639, in download_data
    lchdu  = pf.open(response.url) # this needs to be a URL - not a file
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/io/fits/hdu/hdulist.py", line 150, in fitsopen
    return HDUList.fromfile(name, mode, memmap, save_backup, cache,
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/io/fits/hdu/hdulist.py", line 388, in fromfile
    return cls._readfrom(fileobj=fileobj, mode=mode, memmap=memmap,
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/io/fits/hdu/hdulist.py", line 1039, in _readfrom
    fileobj = _File(fileobj, mode=mode, memmap=memmap, cache=cache)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/utils/decorators.py", line 503, in wrapper
    return function(*args, **kwargs)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/io/fits/file.py", line 147, in __init__
    self.name = download_file(fileobj, cache=cache)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/utils/data.py", line 1012, in download_file
    with shelve.open(urlmapfn) as url2hash:
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/shelve.py", line 243, in open
    return DbfilenameShelf(filename, flag, protocol, writeback)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/shelve.py", line 227, in __init__
    Shelf.__init__(self, dbm.open(filename, flag), protocol, writeback)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/dbm/__init__.py", line 91, in open
    raise error[0]("db type is {0}, but the module is not "
dbm.error: db type is dbm.gnu, but the module is not available

Could you please check this?
The same message appears when running the mpi example....
Cheers
Aurelio
