+++ dkafhjks/data/tesscurl_sector_all_tp.sh
@@ -1,1 +1,0 @@
-#all LC file links
