### 2020-05-12 #2101 @noraeisner
@aureliocarnero, I'm so sorry! One of the files went missing in my last git push (my computer had run out of space and started uploading random files to the cloud and off my computer). I've updated the git (unittests only work with the github version as the files that are needed for the tests are large, and it was recommended to keep the pip version as small as possible). I'm so sorry and thank you so much for you patience with this. That should solve the unittest error. If the mpi version still does not work I will remove this functionality for now, if you agree, as it works on both of my computer and I don't know enough about mpi to diagnose the problem on other operating systems (and it's not key to the code by any means).
Thank you
Nora
