### 2020-05-12 #2101 @aureliocarnero
@noraeisner no problem! That is why we review it, so it ends up being a clean and neat product = ).
I commit errors all the time.

So, I did a fresh installation, using pip3 install tessLATTE, and I get the error I was getting when I was doing the unittest. I can introduce a TIC ID and sectors, but then, this error message appears:

raceback (most recent call last):
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/runpy.py", line 193, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/runpy.py", line 86, in _run_code
    exec(code, run_globals)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/LATTE/__main__.py", line 424, in <module>
    utils.interact_LATTE(tic, indir, syspath, sectors_all, sectors, ra, dec, args)  # the argument of whether to shos the images or not
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/LATTE/LATTEutils.py", line 129, in interact_LATTE
    alltime, allflux, allflux_err, all_md, alltimebinned, allfluxbinned, allx1, allx2, ally1, ally2, alltime12, allfbkg, start_sec, end_sec, in_sec, tessmag, teff, srad = download_data(indir, sectors, tic)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/LATTE/LATTEutils.py", line 2639, in download_data
    lchdu  = pf.open(response.url) # this needs to be a URL - not a file
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/io/fits/hdu/hdulist.py", line 150, in fitsopen
    return HDUList.fromfile(name, mode, memmap, save_backup, cache,
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/io/fits/hdu/hdulist.py", line 388, in fromfile
    return cls._readfrom(fileobj=fileobj, mode=mode, memmap=memmap,
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/io/fits/hdu/hdulist.py", line 1039, in _readfrom
    fileobj = _File(fileobj, mode=mode, memmap=memmap, cache=cache)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/utils/decorators.py", line 503, in wrapper
    return function(*args, **kwargs)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/io/fits/file.py", line 147, in __init__
    self.name = download_file(fileobj, cache=cache)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/site-packages/astropy/utils/data.py", line 1012, in download_file
    with shelve.open(urlmapfn) as url2hash:
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/shelve.py", line 243, in open
    return DbfilenameShelf(filename, flag, protocol, writeback)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/shelve.py", line 227, in __init__
    Shelf.__init__(self, dbm.open(filename, flag), protocol, writeback)
  File "/home/carnero/anaconda3/envs/testLATTE/lib/python3.8/dbm/__init__.py", line 91, in open
    raise error[0]("db type is {0}, but the module is not "
dbm.error: db type is dbm.gnu, but the module is not available
