### 2020-05-12 #2101 @noraeisner
Thank you! I just looked into that error message and it appears to be a problem with the astropy stored caches (https://github.com/astropy/astropy/issues/9607) so I have added the following two lines to the start of the code in LATTteutils.py:

from astropy.utils.data import clear_download_cache
clear_download_cache()

I can't reproduce the same error on my end so I, unfortunately, can't check whether that works but if you could try it that would be amazing (I have updated the git version or you could copy and paste those two lines into the to the top of LATTEutils.py. Thank you
