### 2020-05-12 noraeisner/LATTE#3 @christinahedges
This is great work Nora, thanks for the thorough response. The documentation is much more thorough now!
