### 2020-05-12 #2101 @christinahedges
@noraeisner thank you for addressing my comments, the documentation looks great, and the diagnostic plots are very helpful. This is ready for publication.
