+++ paper/paper.bib
@@ -421,7 +421,7 @@ archivePrefix = {arXiv},
 
 
 @article{2018lightkurve,
-  title={Lightkurve: Kepler and TESS time series analysis in Python},
+  title="{Lightkurve: Kepler and TESS time series analysis in Python}",
   author={Cardoso, Jos{\'e} Vin{\'\i}cius de Miranda and Hedges, Christina and Gully-Santiago, Michael and Saunders, Nicholas and Cody, Ann Marie and Barclay, Thomas and Hall, Oliver and Sagear, Sheila and Turtelboom, Emma and Zhang, Johnny and others},
   journal={Astrophysics Source Code Library},
   year={2018}
