### 2020-05-13 #2101 @xuanxu
@noraeisner I've issued a pull request in your LATTE repo fixing some typos in the paper text. Can you take a look at it?
