### 2020-05-13 #2101 @xuanxu
Thanks @noraeisner! Everything looks ready now.
At this point could you make a new release of this software that includes the changes that have resulted from this review?. Then, please make an archive of the software in Zenodo (or a similar service) For the Zenodo archive make sure that:

Once you do that please report here the new version and the DOI of the archive.
