+++ LATTE/LATTEutils.py
@@ -45,7 +45,6 @@ from LATTE import filters
 from LATTE import LATTEbrew as brew
 
 
-
 '''
 Overview of LATTE scipts:
 
@@ -1531,7 +1530,7 @@ def interact_LATTE_FFI_aperture(tic, indir, sectors_all, sectors, ra, dec, args)
         
         # ----------
         # plot the mean image and plot the extraction apertures on top of it so that one can verify that the used apertures make sense
-        im = np.mean(tpf.flux, axis = 0)
+        im = np.nanmean(tpf.flux, axis = 0)
         # set up the plot - these are stored and one of the images saved in the report      
         fig, ax = plt.subplots(1,2, figsize=(10,5), subplot_kw={'xticks': [], 'yticks': []})
         kwargs = {'interpolation': 'none', 'vmin': im.min(), 'vmax': im.max()}
@@ -2600,7 +2599,7 @@ def download_data(indir,sector, tic, binfac = 5, test = 'no'):
         if len(dwload_link) == 0:
             print ("TIC {} was not observed in Sector(s):   {}. Try again with different sectors.".format(tic, sector))
     
-            print ("\n (Also check that the data file have been downloaded - run code with '--new-data' in the command line)")
+            print ("\n (Also check that the data files have been downloaded - run code with '--new-data' in the command line)")
             raise SystemExit
     
     # define all the empty lists to append to in order to return the data that will be requrides later on in the script
@@ -3115,7 +3114,7 @@ def download_data_FFI(indir, sector, syspath, sectors_all, tic, save = False):
         
         # ----------
         # plot the mean image and plot the extraction apertures on top of it so that one can verify that the used apertures make sense
-        im = np.mean(tpf.flux, axis = 0)
+        im = np.nanmean(tpf.flux, axis = 0)
         # set up the plot - these are stored and one of the images saved in the report      
         fig, ax = plt.subplots(1,2, figsize=(10,5), subplot_kw={'xticks': [], 'yticks': []})
         kwargs = {'interpolation': 'none', 'vmin': im.min(), 'vmax': im.max()}
@@ -3528,7 +3527,7 @@ def download_tpf_lightkurve(indir, transit_list, sector, tic, test = 'no'):
                 print ("In the future, this TIC {} will be observed in sector(s) {}".format(tic, future_sectors))
     
         if len(dwload_link) == 0:
-            print ("\n (Also check that the data file have been downloaded - run code with '--new-data' in the command line)")
+            print ("\n (Also check that the data files have been downloaded - run code with '--new-data' in the command line)")
     
         dwload_link_tp = dwload_link
 
@@ -3600,7 +3599,7 @@ def download_tpf_lightkurve(indir, transit_list, sector, tic, test = 'no'):
     
             # ----------
             # plot the mean image and plot the extraction apertures on top of it so that one can verify that the used apertures make sense
-            im = np.mean(tpf.flux, axis = 0)
+            im = np.nanmean(tpf.flux, axis = 0)
             # set up the plot - these are stored and one of the images saved in the report      
             fig, ax = plt.subplots(1,2, figsize=(10,5), subplot_kw={'xticks': [], 'yticks': []})
             kwargs = {'interpolation': 'none', 'vmin': im.min(), 'vmax': im.max()}
[…]
