+++ paper/paper.bib
@@ -16,12 +16,38 @@
 }
 
 
-@article{eisner19,
-  title={Planet Hunters TESS I: TOI 813, a subgiant hosting a transiting Saturn-sized planet on an 84-day orbit},
-  author={Eisner, NL and Barrag{\'a}n, O and Aigrain, S and Lintott, C and Miller, G and Boyajian, TS and Brice{\~n}o, C and Bryant, EM and Christiansen, JL and Feinstein, AD and others},
-  journal={arXiv preprint arXiv:1909.09094},
-  year={2019}
+@ARTICLE{eisner19,
+       author = {{Eisner}, N.~L. and {Barrag{\'a}n}, O. and {Aigrain}, S. and
+         {Lintott}, C. and {Miller}, G. and {Zicher}, N. and {Boyajian}, T.~S. and
+         {Brice{\~n}o}, C. and {Bryant}, E.~M. and {Christiansen}, J.~L. and
+         {Feinstein}, A.~D. and {Flor-Torres}, L.~M. and {Fridlund}, M. and {Gand
+        olfi}, D. and {Gilbert}, J. and {Guerrero}, N. and {Jenkins}, J.~M. and
+         {Jones}, K. and {Kristiansen}, M.~H. and {Vanderburg}, A. and
+         {Law}, N. and {L{\'o}pez-S{\'a}nchez}, A.~R. and {Mann}, A.~W. and
+         {Safron}, E.~J. and {Schwamb}, M.~E. and {Stassun}, K.~G. and
+         {Osborn}, H.~P. and {Wang}, J. and {Zic}, A. and {Ziegler}, C. and
+         {Barnet}, F. and {Bean}, S.~J. and {Bundy}, D.~M. and {Chetnik}, Z. and
+         {Dawson}, J.~L. and {Garstone}, J. and {Stenner}, A.~G. and
+         {Huten}, M. and {Larish}, S. and {Melanson}, L.~D. and {Mitchell}, T. and
+         {Moore}, C. and {Peltsch}, K. and {Rogers}, D.~J. and {Schuster}, C. and
+         {Smith}, D.~S. and {Simister}, D.~J. and {Tanner}, C. and
+         {Terentev}, I. and {Tsymbal}, A.},
+        title = "{Planet Hunters TESS I: TOI 813, a subgiant hosting a transiting Saturn-sized planet on an 84-day orbit}",
+      journal = {\mnras},
+     keywords = {methods: statistical, planets and satellites: detection, stars: fundamental parameters, stars: individual (TIC-55525572, TOI 813), Astrophysics - Earth and Planetary Astrophysics},
+         year = 2020,
+        month = jan,
+       volume = {494},
+       number = {1},
+        pages = {750-763},
+          doi = {10.1093/mnras/staa138},
+archivePrefix = {arXiv},
+       eprint = {1909.09094},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.494..750E},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
 
 
 @ARTICLE{Christiansen2018,
