### 2020-05-18 #2101 @noraeisner
@xuanxu I have upload the new release, v1.0.0, to Zenodo under the DOI 10.5281/zenodo.3831586. Thank you!
