### 2020-05-18 #2101 @arfon
@noraeisner :wave: hope you are well! Your paper is now accepted into JOSS :zap::rocket::boom:
